requirements.txt
aiogram==3.7.0
fluent.runtime==0.4.0
fluent.syntax==0.19.0
gspread==6.0.2
google-auth==2.35.0
pydantic
pydantic-settings
python-dotenv
